+++ covid-dashboard/src/app/components/content/content.component.html
@@ -5,3 +5,5 @@
   <app-graph></app-graph>
   <app-table></app-table>
 </div>
+
+<p>Hi</p>
